Fixed leak when server is lagging and tidy up poi hook
